move serverconfig and fix imports
