+++ ahnlich/server/tests/server_test.rs
@@ -1,5 +1,4 @@
 use server::ServerConfig;
-
 use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
 use tokio::net::TcpStream;
 use tokio::time::{timeout, Duration};
